visitsikkim_project/requirements.txt
django
pillow
requests
python-dotenv
python-decouple
psycopg2-binary
